requirements.txt
slackbot
jinja2
python-memcached
